refactor(app_runtime): spawn_wizard_shell_window_async を wizard.rs に追加 (SPEC-2077 Phase F3b)
- crates/gwt/src/app_runtime/wizard.rs に spawn_wizard_shell_window_async (36 行) を追加
- mod.rs から該当 method 削除
- これで Phase F (Wizard) 完成: F1 + F2a-1 + F2a-2 + F2b + F3a + F3b 合計 9 method を wizard.rs に集約 (533 行)
- 累計 mod.rs 7445→6011 行 (-1434 行、 god module 19.3% 縮小)

Verification:
- cargo test -p gwt-core -p gwt: 全 pass
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
